Renamed `add` processor to `set` processor.
This name makes more sense, because if a field already exists it overwrites it.
