+++ plugins/ingest/src/test/resources/rest-api-spec/test/ingest/20_crud.yaml
@@ -12,7 +12,7 @@
             "description": "_description",
             "processors": [
               {
-                "add" : {
+                "set" : {
                   "fields" : {
                     "field2": "_value"
                   }
