+++ plugins/ingest/src/test/resources/rest-api-spec/test/ingest/60_mutate.yaml
@@ -12,7 +12,7 @@
             "description": "_description",
             "processors": [
               {
-                "add" : {
+                "set" : {
                   "fields" : {
                     "new_field": "new_value"
                   }
