+++ plugins/ingest/src/test/resources/rest-api-spec/test/ingest/80_simulate.yaml
@@ -12,7 +12,7 @@
             "description": "_description",
             "processors": [
               {
-                "add" : {
+                "set" : {
                   "fields" : {
                     "field2" : "_value"
                   }
@@ -66,7 +66,7 @@
               "description": "_description",
               "processors": [
                 {
-                  "add" : {
+                  "set" : {
                     "fields" : {
                       "field2" : "_value"
                     }
@@ -129,14 +129,14 @@
               "description": "_description",
               "processors": [
                 {
-                  "add" : {
+                  "set" : {
                     "fields" : {
                       "field2" : "_value"
                     }
                   }
                 },
                 {
-                  "add" : {
+                  "set" : {
                     "fields" : {
                       "field3" : "third_val"
                     }
@@ -157,7 +157,7 @@
           }
   - length: { docs: 1 }
   - length: { docs.0.processor_results: 2 }
-  - match: { docs.0.processor_results.0.processor_id: "processor[add]-0" }
+  - match: { docs.0.processor_results.0.processor_id: "processor[set]-0" }
   - is_true: docs.0.processor_results.0.doc.modified
   - length: { docs.0.processor_results.0.doc._source: 2 }
   - match: { docs.0.processor_results.0.doc._source.foo: "bar" }
